latex Guidebook [1 of 2] (trunk only)
     Last fall when Michael added the symset stuff to supersede the old
handling for IBMgraphics and DECgraphics, Guidebook.tex was changed to
support multi-page tables in the output.  But that requires that the
input be processed twice, because it requires feedback stored in
Guidebook.aux and the first pass can't rely on that file being present
or up to date.  This updates the Unix, VMS, and OS/2 makefiles to do
two-pass processing.  (I didn't see latex usage anywhere else, and the
branch version doesn't include the formatting change which needs this.)