wizard mode ^G of shapechangers (trunk only)
     A while back there was a change in how the initial shape for a
shapechanging monster gets chosen during monster creation, and a side-
effect of that lets/makes you choose the shape when #monpolycontrol is
enabled.  But ^G was overriding your choice by forcing the shapechanger
to start out looking like the type of monster that you specified (to ^G,
not to subsequent #monpolycontrol prompting), hence always in its natural
shape.  The intent for ^G was that if asked for a unique monster but got
a doppelganger instead, it would initially look like the requested unique
monster.  Post-3.4.3 code, so no fixes entry needed.