rusty pearl rings (trunk only)
     Recent newsgroup discussion complained about a pearl ring becoming
rusty.  That has no effect on game play but does feel weird.  Rings which
are described by their gems specify a material based on the gem rather
than on the band, so making pearl rings be made out of iron didn't fit.
The substance pearls are made of is the same as in the shells of the
oysters or whatever critters excrete them.  The closest matches available
seem to be mineral and bone; I've gone with bone here.