still held by monster after it moves away (trunk only)
     From a bug report, a stunned monster
moved away from the hero, but remained holding him.  That behavior was
still present in the dev code, although the hero was free to move and
would become unstuck as soon as he did so.  I don't know how many fixes
for this sort of thing have been made, but they evidently haven't been
made in the proper place.  [Perhaps it really ought to be done as a
monster is placed somewhere on the map?]